Add confirmation before delete something

+++ data_lokasi_wisata.php
@@ -220,7 +220,7 @@ if($_SESSION['legitUser'] != 'qwerty'){
 						?>
 						<th><?=$data[strtolower($dakrit['kriteria'])];?></th>
 						<?php endwhile;?>
-                        <th><a href="delete.php?id=<?php echo $data['id']; ?>&item=lokasi"><button class="btn btn-danger">Delete</button></a></th>
+                        <th><a href="delete.php?id=<?php echo $data['id']; ?>&item=lokasi"><button class="btn btn-danger" onclick="return confirmAction()">Delete</button></a></th>
                     </tr>
 
                 <?php $num++; endwhile;?>
@@ -232,3 +232,16 @@ if($_SESSION['legitUser'] != 'qwerty'){
 	</div>
 </body>
 </html>
+
+<script>
+      // The function below will start the confirmation dialog
+      function confirmAction() {
+        let confirmAction = confirm("Anda yakin ingin menghapus data lokasi?");
+        if (confirmAction) {
+			 alert("Berhasil menghapus data lokasi")
+			 return true;
+        } else {
+			return false;
+        }
+      }
+    </script>
